requirements.txt
alembic==1.11.1
aniso8601==9.0.1
attrs==23.1.0
beautifulsoup4==4.9.3
blinker==1.4
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
cryptography==41.0.1
cwe2==2.0.0
dnspython==2.3.0
email-validator==1.1.3
exceptiongroup==1.1.2
Flask==2.0.2
Flask-Dance==5.1.0
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-Minify==0.37
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==2.0.2
gunicorn==20.1.0
htmlmin==0.1.12
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.0.3
joblib==1.3.1
jsmin==3.0.1
jsonschema==4.17.3
lesscpy==0.15.1
Mako==1.2.4
MarkupSafe==2.1.3
neo4j==5.9.0
neo4j-driver==4.4.10
neobolt==1.7.17
neomodel==5.0.1
nltk==3.8.1
numpy==1.25.1
oauthlib==3.2.2
packaging==23.1
pandas==2.0.3
pluggy==1.2.0
ply==3.11
pycparser==2.21
pyOpenSSL==23.2.0
pyrsistent==0.19.3
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2023.3
PyYAML==6.0
rcssmin==1.1.1
regex==2023.6.3
requests==2.26.0
requests-oauthlib==1.3.1
scipy==1.11.1
sentry-sdk==1.26.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==1.4.29
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
URLObject==2.4.3
Werkzeug==2.0.3
WTForms==3.0.1
xmltodict==0.12.0
xxhash==3.2.0
